+++ stac_fastapi/core/stac_fastapi/core/extensions/__init__.py
@@ -1,5 +1,11 @@
 """elasticsearch extensions modifications."""
 
+from .collections_search import CollectionsSearchEndpointExtension
 from .query import Operator, QueryableTypes, QueryExtension
 
-__all__ = ["Operator", "QueryableTypes", "QueryExtension"]
+__all__ = [
+    "Operator",
+    "QueryableTypes",
+    "QueryExtension",
+    "CollectionsSearchEndpointExtension",
+]